requirements.txt
ccxt==4.4.26
python-telegram-bot==21.5
pandas==2.2.3
numpy==1.26.4
ta==0.11.0
python-dotenv==1.0.1
aiohttp==3.10.10
